@@ -1,7 +1,7 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
 
     defaultConfig {
         minSdkVersion 14
@@ -18,14 +18,13 @@ android {
     }
 }
 
-ext.supportLibraryVersion = '27.1.0'
-
 dependencies {
-    implementation "com.android.support:appcompat-v7:${supportLibraryVersion}"
-    implementation "com.android.support:support-v4:${supportLibraryVersion}"
+    implementation 'androidx.appcompat:appcompat:1.0.2'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
     implementation project(':laevatein')
     implementation fileTree(dir: 'libs', include: ['*.jar'])
 
-    implementation 'com.github.hotchemi:permissionsdispatcher:2.2.0'
-    annotationProcessor 'com.github.hotchemi:permissionsdispatcher-processor:2.2.0'
+    implementation 'org.permissionsdispatcher:permissionsdispatcher:4.3.0'
+    annotationProcessor 'org.permissionsdispatcher:permissionsdispatcher-processor:4.3.0'
 }
