@@ -18,8 +18,8 @@ package com.laevatein;
 import android.app.Activity;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.laevatein.ui.PhotoSelectionActivity;
 
