@@ -2,7 +2,7 @@ package com.laevatein.internal.entity;
 
 import android.os.Parcel;
 import android.os.Parcelable;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.laevatein.R;
 
