@@ -18,7 +18,7 @@ package com.laevatein.internal.misc.ui;
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Parcelable;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.View;
 
 /**
