@@ -6,7 +6,7 @@ import com.laevatein.ui.PhotoSelectionActivity;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.View;
