@@ -16,10 +16,10 @@
 package com.laevatein.internal.ui.helper;
 
 import android.database.Cursor;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentActivity;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.widget.CursorAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentActivity;
+import androidx.fragment.app.FragmentManager;
+import androidx.cursoradapter.widget.CursorAdapter;
 import android.util.TypedValue;
 import android.view.LayoutInflater;
 import android.view.View;
