@@ -1,7 +1,7 @@
 package com.laevatein.internal.ui.widget;
 
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 
 public class PhotoDecoration extends RecyclerView.ItemDecoration {
