@@ -15,8 +15,8 @@ import android.os.Build;
 import android.os.Environment;
 import android.os.Handler;
 import android.provider.MediaStore;
-import android.support.media.ExifInterface;
-import android.support.v4.content.FileProvider;
+import androidx.exifinterface.media.ExifInterface;
+import androidx.core.content.FileProvider;
 import android.util.Log;
 
 import java.io.File;
